Add typecasting

+++ smsd/log-event.c
@@ -83,7 +83,7 @@ gboolean eventlog_register(void)
 		0,
 		REG_SZ,
 		(BYTE *)program_name,
-		strlen(program_name) + 1);
+		(DWORD)(strlen(program_name) + 1));
 
 	if (ret != ERROR_SUCCESS) {
 		fprintf(stderr, "Failed to write CategoryMessageFile to registry!\n");
